Implement basic counting algorithm
Now we can handle IVM matview with duplication or DISTINCT. DISTINCT ON
is not supported. When matviews are created, duplication of tpules is
counted and stored into "__ivm_count__" column, which is a special
column for IVM matviews. Users should not use this column name, at
least for matviews, but this is not checked for now.  __ivm_count__
is updated when matviews are maintained incrementally, and a tuple
is removed from the matview when its count becomes zero.

When SELECT is issued for matviews with duplication, each tuple is
returned __ivm_count__ times. generate_series(1, __ivm_count__)
function is called to implement this, but there may be better ways.
In the case of DISTINCT matviews, tuples in these are returned
directly.

Currently, matviews with GROUP BY are not considered yet although
this is not checked when matviews is created.