Prevent race conditions which updates materialized views anormally
In READ COMMITED isolation level, take a lock in early stage to wait for
concurrent transactions which are updating this materialized view.
This is necessary because we have to see changes on base tables committed
in these transactions to prevent anormal update. Also, we have to get
the latest snapshot before computting delta tables. If we could not
know changes of base tables in other concurrent transactions, anormal
update would happen after transaction committed.

In REPEATABLE READ or SERIALIZABLE level, don't wait a lock, and raise
an error immediately to prevent anormal update because changes in
other transaction never been seen in these levels.

These solutions are ugly and there may be better ways, but I commit
this to prevent anormal update, anyway.

Maybe, only way to do better is update materialized views at end of
transactions instead of at end of query during transaction. To
implement this, we will need change log maintenance mechanism.