Fix an error of IVM on aggregation views without GROUP clause
When updating aggregation matviews, it is assumed that group keys
are used to look up tuples to be updated. Therefore, when the view
doesn't have a group clause, this caused an error because wrong syntax,
WHERE ()=() was used in a query.

In this case, there is a only one tuple in the view and we don't
have to use any group key. So, I fixed this using a dummy where
clause which results always in true, like WHERE (1)=(1).