Fix an error of avg aggregate in IVM calculation
If an avg result in materialized views is updated incrementally by
using the old avg value, errors in avg are accumulated and the value
gets wrong eventually. To prevent this, contain sum and count values
in views and use them to calculate avg instead of using old avg values.
sum and count can be updated incrementally without accumlating errors.