Add support for min/max aggregation function (#19)
If new tuples are inserted and new values are small (for min) or
large (for max), view shoudl be upated with the new values. Otherwise,
old values remains.

In the case of deletion, this is more complicated. If deleted values
are current min or max, we have to recalc new min or max values for
affected groups and update the view with these value.