Fix an error of min/max aggregetions withoug GROUP BY clause
This situation was not considered whithin the previous commit.
The regression tests for this view definition are added.

However, additional regression tests uncovered another bug that
plan caches for queries to update min/max values don't work because
only one cache is shared among all matviews. To avoid this stupid
bug, the codes to use this plan caches are commented out. This will
be fixed soon by the following commit.