Fix incorrect avg value when types other than int are used
For calculating delta of arg, matviews store sum and count values
internally, but the internal sum value was incorrect because the
sum aggregate function for int type argument was always used. This
was due to wrong setup of dummy argument passed to ParseFuncOrColumn
before looking up OID of the function. In this fix the correct dummy
arguments re made from OIDs of the original avg function's arguments.