Allow to use aggregate functions for interval type (#27)
Aggregate functions for interval type were not considered
and errors occurred when we tried to update matviews due to
add operation with zero value which are not permitted for
interval type. Now addition with zero is not assumed and
aggregate for interval works well. In passing, codes for
updating aggregate values were refactored.

issue #25