Use tuplestore and ENR instead of temptable
In the previous implementation, temporary tables were used during maintenance 
of materialized views. Tow delta tables (as tuple sets to be removed and inserted)
were created and dropped per a statement due to immediate maintenance.

This was problematic for following reasons:

- Frequent create/drop may bloat system catalogue tables. Also it may take many
  locks and cause "out of shared memory" error.
 (https://www.postgresql.org/message-id/1577564109604-0.post%40n3.nabble.com )

- Temptables can not be used in prepared transactions.
- Creating new tables per statement prevents plan cache.
- It may be security problems, as well as performance problems

This was resolved by using tuplestore and ENRs rather than temptables.