Fix to not use generate_series when views are queried
In the previous implementation, multiplicity of each tuple was stored
in ivm_count column in views. When SELECT was issued for views with
duplicate, the view was replaced with a subquery in which each tuple
was joined with generate_series function in order to output tuples
of the number of ivm_count.

This was problematic for following reasons:

- The overhead was huge. When almost of tuples in a view were selected,
  it took much longer time than the original query. This lost the meaning
  of materialized views.

- Optimizer could not estimate row numbers correctly because this had to
  know ivm_count values stored in tuples.

- System columns of materialized views like cmin, xmin, xmax could not
  be used because a view was replaced with a subquery.

To resolve this, the new implementation doen't store multiplicities
for views with tuple duplicates, and doesn't use generate_series
when SELECT query is issued for such views.

Note that we still have to use ivm_count for supporting DISTINCT and
aggregates.

Design:

Alothough views doesn't have ivm_count, multiplicities for old delta
and new delta are calculated and the count is contained in each table.

Each tuple in the old delta is applied using ctid and LIMIT function, like:

 DELETE FROM matviewname WHERE ctid IN (
    SELECT ctid FROM matviewname
    WHERE (key1, key2, ..., keyn) = ($1, $2, ..., $n) LIMIT $(n+1));

The new delta is applied using generate_seriese, like:

 INSERT INTO matviewname (c1, c2, ...)
   SELECT c1,c2,... FROM (
      SELECT diff.*, generate_series(1, diff.__ivm_count__)
      FROM new_delta AS diff) AS v