Add outer-join support in IVM
In case of outer-joins, additionally to deltas which occur in inner-join
case we need additional delete or insert of dangling tuples, that is,
null-extended tuples generated when join-condition doesn't match.

This implementation basically based on the algorithm of Larson & Zhou
(2007) [1]. Before view maintenances, jointree is analysed and "view
maintenance graph" is generated, which represents which tuples in the
views are affected when a table is modified. Tuples on which the modified
table is not null-extended are directly affected by this change. Such
affects are calculated similarly to inner-joins. On the other hand,
dangling tuples generated by joining (directly) affected tuples can be
indirectly affected. This means that we may need to delete dangling
tuples when any tuples are inserted to a table, as well as to insert
dangling tuples when tuples are deleted from a table.

[1] Efficient Maintenance of Materialized Outer-Join Views (Larson & Zhou, 2007)
https://ieeexplore.ieee.org/document/4221654

Currently, we have following restrictions:

- outer join view's targetlist must contain attributes used in join conditions
- outer join view's targetlist cannot contain non-strict functions
- outer join supports only simple equijoin
- outer join view's WHERE clause cannot contain non null-rejecting predicates
- aggregate is not supported with outer join
- subquery is not supported with outer join