Add comment for SetTransitionTablePreserved and related (#106)
Also, remove the code which prolong transition tables automatically
for IVM triggers. This code is not necessary if we call the function
explicitly. This was left accidentally.

Although I am not sure if there is any use case of this function
other than IVM, I chose to use the function considering to avoid
to embed IVM specific code in trigger.c.

Per comments from Tatsuo Ishii in pgsql-hackers.