Add support for simple CTEs
Simple CTEs which does not contain aggregates or DISTINCT are now
supported similarly to simple sub-queries.

Before a view is maintained, all CTEs are converted to corresponding
subqueries to enable to treat CTEs as same as subqueries. For this
end, inline_cte in optimizer/plan/subselect.c was export to public.

Related issue #8