Use weaker lock on matview if possible (#119)
If the view has only one base table in this query, we can use
RowExclusiveLock instead of AccessExclusiveLock, because we don't
need to wait other concurrent transaction's result in order to
maintain the view in this case. When the same row in the view is
affected due to concurrent maintenances, a row level lock will
protect it.