Do not create index when creating IMMV WITH NO DATA
It is to avoid an error when restoring pg_dump results. The script
output by pg_dump includes CREATE INCREMENTAL MATERIALIZED VIEW WITH
NO DATA followed by CREATE UNIQUE INDEX. Therefore, if the former
creates a unique index, the latter to create the same index with the
identical name ends with an error. Also, the index may be created
by REFRESH WITH DATA only when the same one is not created yet.