Fix bugs of IVM that occur when column names are specified in aggregate views
The names of hidden columns added for aggregate views are derived from column
names specified in CREATE INCREMENTAL MATERIALIZED VIEW, but when the number
was less than the length of the actual target list, segmentation fault occurred.
Furthermore, when the number of specified columns is more than the target list,
it overrode hidden column names and it caused a failure of incremental maintenance
of an aggregate view.

To fix then, check the length of the specified column name list not to access
invalid area, and also prevent from overriding hidden column names.